+++ test/libsolidity/SolidityNameAndTypeResolution.cpp
@@ -1115,7 +1115,7 @@ BOOST_AUTO_TEST_CASE(function_modifier_double_invocation)
 			modifier mod(uint a) { if (a > 0) _; }
 		}
 	)";
-	CHECK_ERROR(text, DeclarationError, "Modifier already used for this function");
+	success(text);
 }
 
 BOOST_AUTO_TEST_CASE(base_constructor_double_invocation)