+++ Changelog.md
@@ -14,6 +14,7 @@ Features:
 Bugfixes:
  * Type Checker: Fix invalid "specify storage keyword" warning for reference members of structs.
  * Type Checker: Mark modifiers as internal.
+ * Type Checker: Re-allow multiple mentions of the same modifier per function.
 
 
 ### 0.4.13 (2017-07-06)